refactor: altera diretório dos teste para jetstream
