+++ .env.testing
@@ -7,7 +7,6 @@ LOG_CHANNEL=stack
 
 DB_CONNECTION=sqlite
 DB_DATABASE=database/database.sqlite
-DB_FOREIGN_KEYS=true
 
 BROADCAST_DRIVER=log
 CACHE_DRIVER=array
